refactor(settings): 拆出云同步提供方定义与设置弹窗助手
- 新增 settings/：settingsDialogs（设置域 confirm/choice/prompt 统一入口）、sync/syncProviders（4 个云同步提供方的类型/默认表单/PROVIDERS 注册表）
- SettingsModal.tsx 2708 -> 2486 行
