+++ frontend/scripts/style-baseline.json
@@ -1,5 +1,5 @@
 {
-  "hexColor": 34,
+  "hexColor": 32,
   "numericFontSize": 20,
   "magicZIndex": 4
 }
